python-http/requirements.txt
anyio==3.6.2
asgiref==3.6.0
asyncio==3.4.3
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.95.0
googleapis-common-protos==1.56.2
grpcio==1.53.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.1
opentelemetry-api==1.17.0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-jaeger-proto-grpc==1.17.0
opentelemetry-exporter-jaeger-thrift==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
protobuf==3.20.3
pydantic==1.10.7
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
thrift==0.16.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wrapt==1.15.0
zipp==3.15.0
